backend mukodik

+++ backend/database/migrations/2026_01_30_111800_create_items_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
 
             $table->id('cikk_szam');
             $table->string('elnevezes', 50);
-            $table->tinyInteger('akt_keszlet');
+            $table->integer('akt_keszlet');
             $table->decimal('egyseg_ar', 10, 2);
 
             $table->timestamps();
